+++ rio/components/button.py
@@ -155,6 +155,7 @@ class Button(Component):
                 children.append(
                     rio.Text(
                         text,
+                        justify="center",
                         margin_x=CHILD_MARGIN_X if n_children == 1 else None,
                         margin_y=CHILD_MARGIN_Y if n_children == 1 else None,
                         style=(